View content of email, apply strong for Unread, allow starring and apply Kaminari

+++ app/presenters/users/index_presenter.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 class Users::IndexPresenter < BasePresenter
   presents :users
 
   def change_language(locale, translation_path)
-    is_language_in(locale, translation_path) do
+    language_in?(locale, translation_path) do
       h.link_to translation_path, root_path(locale: User::CHINESE_CODE), class: 'link-color'
     end
   end
@@ -23,6 +25,10 @@ class Users::IndexPresenter < BasePresenter
     h.link_to translation_path, draft_path(locale: I18n.locale), remote: true
   end
 
+  def starred(translation_path)
+    h.link_to translation_path, starred_path(locale: I18n.locale), remote: true
+  end
+
   def delete(translation_path)
     h.link_to translation_path, delete_path(locale: I18n.locale), remote: true
   end
@@ -32,14 +38,12 @@ class Users::IndexPresenter < BasePresenter
   end
 
   def sign_out(translation_path)
-    if user_signed_in?
-      h.link_to translation_path, destroy_user_session_path, method: :delete, class: 'link-color'
-    end
+    h.link_to translation_path, destroy_user_session_path, method: :delete, class: 'link-color' if user_signed_in?
   end
 
   private
 
-  def is_language_in(language, translation_path)
+  def language_in?(language, translation_path)
     if language == User::ENGLISH_CODE
       yield
     else
